xml_uploader-main/requirements.txt
Django==5.0.4
djangorestframework==3.14.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
django-extensions-4.1